+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Unwieghted Kappa (via `kap`) is availible and is also returned by `metrics`. 
 * Detection prevalence and balanced accuracy were added. 
+* `roc_curve` is a tidy method for getting the points on an ROC curve. 
 
 # `yardstick` 0.0.1